added serviceConfig file and made changes to exec function
